requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
pytest
python-dotenv
